% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covid19Il6JakInhibitorsSccs.R
\docType{package}
\name{Covid19Il6JakInhibitorsSccs-package}
\alias{Covid19Il6JakInhibitorsSccs}
\alias{Covid19Il6JakInhibitorsSccs-package}
\title{Covid19Il6JakInhibitorsSccs: Self-Controlled Case Series Analysis of the Safety of IL-6 and JAK Inhibitors}
\description{
IL-6 and JAK inhibitors are being considered for use in treatment and prophylaxis of COVID-19 in rapid clinical trials across the world. However, the full safety profiles of these drugs is unknown, and the current trials are unlikely to be powered or have sufficient follow-up time to evaluate most safety outcomes. The aim of this OHDSI study is to use existing retrospective data to evaluate the safety of IL-6 and JAK inhibitors, using the self-controlled case series (SCCS) design.
}
\author{
\strong{Maintainer}: Martijn Schuemie \email{schuemie@ohdsi.org}

}
\keyword{internal}
