% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fix.R
\name{deleteHoiFiles}
\alias{deleteHoiFiles}
\title{Delete health outcomes of interest result files}
\usage{
deleteHoiFiles(outputFolder)
}
\arguments{
\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes
(/). Do not use a folder on a network drive since this greatly impacts
performance.}
}
\description{
Delete health outcomes of interest result files
}
