% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyApps.R
\name{launchEvidenceExplorer}
\alias{launchEvidenceExplorer}
\title{Launch the Evidence Explorer Shiny app}
\usage{
launchEvidenceExplorer(dataFolder, blind = TRUE, launch.browser = FALSE)
}
\arguments{
\item{blind}{Should the user be blinded to the main results?}

\item{launch.browser}{Should the app be launched in your default browser, or in a Shiny window.
Note: copying to clipboard will not work in a Shiny window.}

\item{shinyDataFolder}{A folder where the Shiny data are stored. Use
the \code{\link{prepareForEvidenceExplorer}} function to generate these files.}
}
\description{
Launch the Evidence Explorer Shiny app
}
\details{
Launches a Shiny app that allows the user to explore the study results.
}
