% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyApps.R
\name{prepareForEvidenceExplorer}
\alias{prepareForEvidenceExplorer}
\title{Prepare results for the Evidence Explorer Shiny app.}
\usage{
prepareForEvidenceExplorer(resultsFolder, shinyDataFolder)
}
\arguments{
\item{resultsFolder}{Folder where the exported zip files from one or more databases are stored.}

\item{shinyDataFolder}{Folder where the data files for the Shiny app will be written.}
}
\description{
Prepare results for the Evidence Explorer Shiny app.
}
